requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
bcrypt==5.0.0
billiard==4.2.2
Brotli==1.1.0
build==1.4.0
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
cssselect2==0.8.0
dateparser==1.2.2
decorator==5.2.1
defusedxml==0.7.1
ecdsa==0.19.1
executing==2.2.1
fastapi==0.116.1
fonttools==4.60.1
fpdf2==2.8.4
google-api-core==2.29.0
google-api-python-client==2.149.0
google-auth==2.49.0.dev0
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==2.18.2
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.10
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
kombu==5.5.4
libsql-experimental==0.0.55
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
pip-tools==7.5.2
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.2
pyphen==0.17.2
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.22
pytz==2025.2
redis==5.2.1
regex==2025.8.29
reportlab==4.4.9
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlalchemy-libsql==0.2.0
stack-data==0.6.3
starlette==0.47.3
tinycss2==1.4.0
tinyhtml5==2.0.0
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
upstash-redis==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zopfli==0.2.3.post1
